pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cislunar-optical-nav"
version = "0.1.0"
description = "Cislunar optical navigation and midcourse correction analysis"
authors = [{name = "Varsha Krishnakumar"}]
dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
    "opencv-python",
]

[project.optional-dependencies]
high-fidelity = [
    "spiceypy",
]
fast = [
    "numba",
]
yaml-configs = [
    "pyyaml",
]
parquet = [
    "pyarrow",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["cisopt*", "dynamics*", "nav*", "guidance*", "mc*", "orbits*", "utils*", "cv*", "visualization*"]

